Added function to create clean municipalities
The function cleans the geo information of the municipalities.
NOTE: When I refer to municipalities, I mean municipality association
(Gemeindeverband) instead of municipality directly (Gemeinde). The reason
is that the lower level of direct municipalities (Gemeinde) could (partially)
result in too few observations for the estimation in the next step.