Handling of duplicates and adding variables for regression
Handled duplicates in apartment rents which are not captured by the algorithm
through the combination of obid and spell since the maximum spell appears
multiple times for the exceptions listed (same approach as for apartment
sales where this happends as well).

Added addtional variables which are needed in the next step, the estimation.
This includes different timing variables as well as IDs as the combination
of the regional and time level. I also added average prices for these IDs.