Removed "betreut" from independents
Removed the independent variable "betreut" from the set of variables for apartment sales (WK).

The reason is that this variable was only used by ImmoScout up to 2016 and was discontinued afterwards. This leads to the problem that the factor variable only has one value after 2016 (= 0) and hence, the regression cancels.