Generalized the function
It now can take the regional type and adds housing type to the grid specification.

The regional type allows the application of the function not just to grids but also to the aggregated data.

The housing type needs only be added to the grid version because here are the housing types pooled together and exported into one file/ sheet. For the individual (non-grid) data where are separated sheets for each type.