Create creating_export_workbooks.R
Function to create empty Excel workbooks for the final datasets that are delivered to the user.

Needed because in order to append a workbook, it needs to exist. However, creating the workbook within the export functions does not work because if certain steps are not re-run (i.e. skipped) the workbook is not created while another step tries to write to it.

Closes #20