No quarterly effects in regions
The quarterly effects for the regional estimation have been missing because "quarter" was not part of the loop.

I change the structure such that "quarter" is now part of the loop. It also changed the output of the most functions as it is not a list (containing year and quarter values) similar to the functions for regional effects changes.

Closes #17