Fixed typo in reshaping data
Due to a typo in the reshaped object, the data for WM on the grid-level got copied to the CI cells as well (i.e. duplicating WM values and neglecting CI values). This only impacted the final GRIDS file and the sheets Grids_RegionEff_yearly and Grids_RegionEff_Change_yearly. The other exported files were not affected.

This error is the reason for having V14.1.